LPS-105873 Move to downloadNode
